Extract on_failed_login from __init__
